Use results stream to send error notifications. (#519)
This is cleaner, and closer to the end goal state of only having one
loader architecture. It also solves a bug on windows, wherein
autocompletion fires before error resolution, which prefills the summary
but doesn't send errors over the notification manager (but does emit the
DirectivesResult over the stream). Then when error resolution comes in,
the angular driver thinks its getting a repeat result and it doesn't
bother resending errors.