Lexically and unintelligently accept `===` by replacing it with `== `. (#535)
This is not the cleanest solution from a "language theory" perspective.
I did try extending [parseEqualityExpression] to accept the new
operator, but it isn't particularly clean since equality operators have
a few micro details (allowing super==, disallowing x == y == z). Also,
since === has the same precedence as ==, they must share a loop which
means copying the lower implementation, which seemed like too much
coupling, on top of the fact that its already extending something that
will have a limited lifetime.

By replacing === with ==, we may see some strange behavior. For
instance, 'x === y === z' will highlight two characters of the second
'===' when it complains about doing equality with an equality.

But the important thing is that it doesn't matter if we butcher string
literals, and there is no other type of token we can be replacing here.
And the semantics of == vs === should be the same statically for users.