Add support for `angular:` toplevel in analysis_options.yaml.
Also bring back the merge code that was there before for when I handled
the case of configuring `plugins: angular` as well as
`plugins:angular_analyzer_plugin`. That code didn't make sense when that
would result in two plugins loaded at once, however, now it is
reasonable (albeit deprecated) to have both the angular top level, and
the `options: angular(analyzer_plugin)?` section configured at the same
time.

Its not critical we handle it perfectly, but it is important that we
handle it consistently until that option is removed.