Split wildcards into a separate class of selector. (#541)
* Split wildcards into a separate class of selector.

Note: Looking at this code, its actually working incorrectly. It should
require a value, and [x*=y] should match attributes with name [x] when
it has a value containing "y". However, I'll fix the logic in a different
CL.

* Remove test_availDirective_wildcard_attribute_begin, which is invalid

`[x*]` is not valid syntax. For `[x*=y]`, we don't suggest `x` because
we assume that the need to specify `y` is non-obvious and suggesting `x`
may actually be confusing. For `[x]`, we have other tests already.