Format with latest dartfmt (#589)
Format with latest dartfmt, handle dart-lang/dart_style#711

The string within the string interpolation gets formatted due to that bug in a way that changes the semantics (https://github.com/dart-lang/dart_style/issues/711). Change the semantics such that the formatter does not attempt to format that string but in a way that still passes the tests.