Fix (click) issues: Find events without @DomName, removed from dart:html.
It looks like DomName can be inferred by lowercasing the accessor name.
This is not reliable for inputs, but for events it seems to be fully
reliable.

Make the changes to the mock SDK, these changes have also been tested
against the latest real SDK. We need integration tests for this type of
thing.

Fix #591