Validate `[attr.foo.if]`: check type bool, and `[attr.foo]` exists. (#624)
As a bonus, make `[attr.foo.if]` click-thru navigable to `[attr.foo]`.

Did some cleanup of the way that these binding type names are cut off.
The diff is ugly but it's mostly just a simple rename. Previous naming
was not clear, and the change required to lop off the 'if' from the name
was not easy to clearly add in.